Always save dependency cache under new key
Under current (default) configuration, CVE database is refreshed every 4
hours. Consequently, a new version of the cache should be saved every 4
hours as well to prevent multiple builds from using the same expired
cache.
By using the `github.run_id` which is unique for each run within a repo,
it means each run will generate its own version of the cache. In the
case were more than 1 run happen for a 4-hour period, then all but the
first run will restore the previous cache and save it unmodified under a
new key. While creating copies of an unmodified cache is wasteful, it
seems to be the most efficient way to keep the cache up-to-date and
reduce cache refresh to a minimum.
In comparison, the previous behaviour implies that while the content of
the gradle files is not modified, an exactly matching cache key is
found. This means that the cache is rarely saved (exact matches do not
save cache, ie. cache is immutable) and that many runs are performed
against the same outdated version of the cache which must be refreshed
as part of the run.