Combine prefixed scopes and roles as authorities
Removal of roles from authorities in 081b5bb was a mistake, this change reverts authorities to being the concatenation of both scopes and roles.
However, it is not correct for roles and scopes to not distinguishable in authorities.
Spring addresses this by prefixing scopes with `SCOPE_` and roles with `ROLE_`.
This change follows this practice.